dts: cm5: Define phy irq pin as input
The support for phy interrupts on BCM54213PE introduced in commit
720694e40c6d ("net: phy: broadcom: Handle irqs on BCM54213PE") relies
on a properly configured interrupt pin. Even though it seems to be
working without specifially configuring the attached gpio as input,
it often was not the case on cm5io board when other irq than 167 where
mapped (see [1]).

Fix this and configure the interrupt pin on gpio 37 as an input.

[1] https://github.com/raspberrypi/linux/issues/6953

Signed-off-by: Nicolai Buchwitz <nb@tipi-net.de>